First steps - the idea and conception
We originally planned to present our project as a Robo-Cat controlled via a Bluetooth module, but we faced technical issues: we did not have enough servo motors for the assembly (the required number was unavailable because we were not given permission to use them) and our team had a limited budget.

Therefore, during the midterm exam , we created a new project "Obstacle Avoiding Robot" in Tinkercad, only cause of the task's criteria. And then, we remembered about it and decided to use that design as the basis for our main project. We discussed this idea with an acquaintance from an engineering university in Korea, and he encouraged us to develop a **Line Follower with Obstacle Avoiding Robot** due to its simplicity in mechanics and control.