Second step - Components
Our team bought all the needed components for the machine that we couldn't find in the Lab: 
like new motors, own arduino, battery, IR-sensors, L298 Motor Driver. 

The chassis for machine we found in last year's students' components in the lab (also we've added the extra part for one of the main boards, so all the components would be fitted). 
So the Hardware Assembling could be started.