Fourth step - half of Hardware
Connecting part:

We glued the last components and added the switch button. Connected the first wires from Arduino to L298 motor driver and breadboard. Also 2 chassises were finally connected with each other by the wiring the wheels' motors with the driver motor.

Then the wires from battery holder was connected to the switch button and connected the button to the driver motor.