Fifth step - finished wiring
We added the ultrasonic sensor and connected with servo-motor. Also we put the wires of it to the breadboard.

After some time (it took us about 3-4 days to connect the wires) we found the IR sensors in the shop, cause we thought we could find them in Lab, but we didn't :(

P.S. we also connected those sensors to the breadboard and that's how we finalized that part of project.